separated fts into files
