+++ src/error_utils.c
@@ -1,48 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils.c                                            :+:      :+:    :+:   */
+/*   error_utils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: linhnguy <linhnguy@hive.student.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/02 17:38:41 by linhnguy          #+#    #+#             */
-/*   Updated: 2024/06/04 13:47:09 by linhnguy         ###   ########.fr       */
+/*   Created: 2024/06/05 11:48:53 by linhnguy          #+#    #+#             */
+/*   Updated: 2024/06/05 12:01:18 by linhnguy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
-
-int	array_len(char **a)
-{
-	int	i;
-
-	i = 0;
-	while (a[i])
-		i++;
-	return (i);
-}
-
-void	remove_string(char **env, int index)
-{
-	free(env[index]);
-	while (env[index + 1])
-	{
-		env[index] = env[index + 1];
-		index++;
-	}
-	env[index] = NULL;
-}
-
-_Bool	ft_isdigit_str(char *str)
-{
-	while (*str)
-	{
-		if (*str < '0' && *str > '9')
-			return (0);
-		str++;
-	}
-	return (1);
-}
 
 void	set_error_and_print(t_pipex *data, int error, char *msg)
 {
@@ -69,21 +37,4 @@ char	*set_error_return(t_pipex *data, int error, char *msg)
 	data->exitcode = error;
 	ft_printf(2, "%s\n", msg);
 	return (NULL);
-}
-
-void	*ft_memcat(void *dst, const void *src)
-{
-	char		*d;
-	const char	*s;
-
-	d = (char *)dst;
-	s = (const char *)src;
-	if (!dst || !src)
-		return (dst);
-	while (*d)
-		d++;
-	while (*s != '\0')
-		*d++ = *s++;
-	*d = '\0';
-	return (dst);
 }
