+++ src/string_utils.c
@@ -1,0 +1,45 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   string_utils.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: linhnguy <linhnguy@hive.student.fi>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/06/05 11:55:53 by linhnguy          #+#    #+#             */
+/*   Updated: 2024/06/05 12:36:58 by linhnguy         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/minishell.h"
+
+int	array_len(char **a)
+{
+	int	i;
+
+	i = 0;
+	while (a[i])
+		i++;
+	return (i);
+}
+
+_Bool	ft_isdigit_str(char *str)
+{
+	while (*str)
+	{
+		if (*str < '0' && *str > '9')
+			return (0);
+		str++;
+	}
+	return (1);
+}
+
+void	remove_string(char **env, int index)
+{
+	free(env[index]);
+	while (env[index + 1])
+	{
+		env[index] = env[index + 1];
+		index++;
+	}
+	env[index] = NULL;
+}
